revised bfa-snn to be compliant w/ new format
