+++ exhibits/bfa_snn/bfasnn_model.py
@@ -1,62 +1,26 @@
-from ngcsimlib.controller import Controller
 from ngclearn.utils.io_utils import makedir
 from ngclearn.utils.viz.raster import create_raster_plot
 from ngclearn.utils.viz.synapse_plot import visualize
 from jax import numpy as jnp, random, jit
 import time
-from ngclearn.utils.model_utils import softmax
 import sys
 
-def tensorstats(tensor):
-    if tensor is not None:
-        _tensor = jnp.asarray(tensor)
-        return {
-            'mean': _tensor.mean(),
-            'std': _tensor.std(),
-            'mag': jnp.abs(_tensor).max(),
-            'min': _tensor.min(),
-            'max': _tensor.max(),
-        }
-    else:
-        return {
-            'mean': None,
-            'std': None,
-            'mag': None,
-            'min': None,
-            'max': None,
-        }
+from ngclearn.utils.model_utils import scanner
+from ngcsimlib.compilers import compile_command, wrap_command
+from ngcsimlib.context import Context
+from ngcsimlib.commands import Command
+from ngcsimlib.operations import summation
 
-def component_stats(component):
-    maxlen = max(len(c) for c in component.compartments.keys()) + 5
-    lines = f"[{component.__class__.__name__}] {component.name}\n"
-    for comp_name, comp_value in component.compartments.items():
-        stats = tensorstats(comp_value)
-        line = [f"{k}: {v}" for k, v in stats.items()]
-        line = ", ".join(line)
-        lines += f"  {f'({comp_name})'.ljust(maxlen)}{line}\n"
-    if hasattr(component, "weights"):
-        stats = tensorstats(getattr(component, "weights"))
-        line = [f"{k}: {v}" for k, v in stats.items()]
-        line = ", ".join(line)
-        lines += f"  {f'(weights)'.ljust(maxlen)}{line}\n"
-    if hasattr(component, "biases"):
-        stats = tensorstats(getattr(component, "biases"))
-        line = [f"{k}: {v}" for k, v in stats.items()]
-        line = ", ".join(line)
-        lines += f"  {f'(biases)'.ljust(maxlen)}{line}\n"
-    return lines
+
+from ngclearn.utils.model_utils import softmax
+from ngclearn.components import GaussianErrorCell, SLIFCell, BernoulliCell, HebbianSynapse
+
 
 ## SNN model co-routines
-def load_model(model_dir, exp_dir="exp", model_name="snn_stdp", dt=3, T=100):
+def load_model(exp_dir="exp", dt=3, T=100):
     _key = random.PRNGKey(time.time_ns())
-    ## load circuit from disk
-    circuit = Controller()
-    circuit.load_from_dir(directory=model_dir)
-
     model = BFA_SNN(_key, in_dim=1, out_dim=1, save_init=False, dt=dt, T=T)
-    model.circuit = circuit
-    model.exp_dir = exp_dir
-    model.model_dir = "{}/{}/custom".format(exp_dir, model_name)
+    model.load_from_disk(exp_dir)
     return model
 
 @jit
@@ -91,10 +55,12 @@ class BFA_SNN():
         save_init: save model at initialization/first configuration time (Default: True)
     """
     # Define Functions
-    def __init__(self, dkey, in_dim, out_dim, hid_dim=1024, T=100, dt=0.25,
-                 tau_m=20., exp_dir="exp", model_name="snn_bfa", save_init=True, **kwargs):
+    def __init__(self, dkey, in_dim=1, out_dim=1, hid_dim=1024, T=100, dt=0.25,
+                 tau_m=20., exp_dir="exp", model_name="snn_bfa", loadDir=None, **kwargs):
         self.exp_dir = exp_dir
+        self.model_name = model_name
         makedir(exp_dir)
+        #makedir("{}/{}".format(exp_dir, model_name))
         makedir(exp_dir + "/filters")
         makedir(exp_dir + "/raster")
 
@@ -119,98 +85,145 @@ class BFA_SNN():
 
         ################################################################################
         ## Create/configure model and simulation object
-        circuit = Controller()
+        # circuit = Controller()
 
-        z0 = circuit.add_component("bernoulli", name="z0", n_units=in_dim, key=subkeys[0])
-        W1 = circuit.add_component("hebbian", name="W1", shape=(in_dim, hid_dim),
-                                   eta=1., wInit=weightInit, bInit=biasInit,
-                                   signVal=-1., optim_type=optim_type, w_bound=0.,
-                                   pre_wght=1., post_wght=eta1_w, is_nonnegative=False,
-                                   key=subkeys[1])
-        z1 = circuit.add_component("SLIF", name="z1", n_units=hid_dim, tau_m=tau_m, R_m=R_m,
-                                   thr=v_thr, inhibit_R=0., sticky_spikes=True,
-                                   refract_T=refract_T, thrGain=0., thrLeak=0.,
-                                   thr_jitter=0., key=subkeys[2])
-        W2 = circuit.add_component("hebbian", name="W2", shape=(hid_dim, out_dim),
-                                   eta=1., wInit=weightInit, bInit=biasInit,
-                                   signVal=-1., optim_type=optim_type, w_bound=0.,
-                                   pre_wght=1., post_wght=eta2_w, is_nonnegative=False,
-                                   key=subkeys[3])
-        z2 = circuit.add_component("SLIF", name="z2", n_units=out_dim, tau_m=tau_m, R_m=R_m,
-                                   thr=v_thr, inhibit_R=0., sticky_spikes=True,
-                                   refract_T=refract_T, thrGain=0., thrLeak=0.,
-                                   thr_jitter=0., key=subkeys[4])
-        e2 = circuit.add_component("error", name="e2", n_units=out_dim)
-        E2 = circuit.add_component("hebbian", name="E2", shape=(out_dim, hid_dim),
-                                   eta=0., wInit=weightInit, bInit=None, w_bound=0.,
-                                   is_nonnegative=False, key=subkeys[5])
-        d1 = circuit.add_component("error", name="d1", n_units=hid_dim)
+        if loadDir is not None:
+            ## build from disk
+            self.load_from_disk(loadDir)
+        else:
+            with Context("Circuit") as self.circuit:
+                self.z0 = BernoulliCell(name="z0", n_units=in_dim, key=subkeys[0])
+                self.W1 = HebbianSynapse(name="W1", shape=(in_dim, hid_dim),
+                                       eta=1., wInit=weightInit, bInit=biasInit,
+                                       signVal=-1., optim_type=optim_type, w_bound=0.,
+                                       pre_wght=1., post_wght=eta1_w, is_nonnegative=False,
+                                       key=subkeys[1])
+                self.z1 = SLIFCell(name="z1", n_units=hid_dim, tau_m=tau_m, R_m=R_m,
+                                       thr=v_thr, inhibit_R=0., sticky_spikes=True,
+                                       refract_T=refract_T, thrGain=0., thrLeak=0.,
+                                       thr_jitter=0., key=subkeys[2])
+                self.W2 = HebbianSynapse(name="W2", shape=(hid_dim, out_dim),
+                                       eta=1., wInit=weightInit, bInit=biasInit,
+                                       signVal=-1., optim_type=optim_type, w_bound=0.,
+                                       pre_wght=1., post_wght=eta2_w, is_nonnegative=False,
+                                       key=subkeys[3])
+                self.z2 = SLIFCell(name="z2", n_units=out_dim, tau_m=tau_m, R_m=R_m,
+                                       thr=v_thr, inhibit_R=0., sticky_spikes=True,
+                                       refract_T=refract_T, thrGain=0., thrLeak=0.,
+                                       thr_jitter=0., key=subkeys[4])
+                self.e2 = GaussianErrorCell(name="e2", n_units=out_dim)
+                self.E2 = HebbianSynapse(name="E2", shape=(out_dim, hid_dim),
+                                       eta=0., wInit=weightInit, bInit=None, w_bound=0.,
+                                       is_nonnegative=False, key=subkeys[5])
+                self.d1 = GaussianErrorCell(name="d1", n_units=hid_dim)
 
-        ## wire up z0 to z1 via W1
-        circuit.connect(z0.name, z0.outputCompartmentName(), W1.name, W1.inputCompartmentName())
-        circuit.connect(W1.name, W1.outputCompartmentName(), z1.name, z1.inputCompartmentName())
 
-        circuit.connect(z1.name, z1.outputCompartmentName(), W2.name, W2.inputCompartmentName())
-        circuit.connect(W2.name, W2.outputCompartmentName(), z2.name, z2.inputCompartmentName())
+                ## wire up z0 to z1 via W1
+                self.W1.inputs << self.z0.outputs
+                self.z1.j << self.W1.outputs
 
-        circuit.connect(e2.name, e2.derivMeanName(), E2.name, E2.inputCompartmentName())
-        circuit.connect(E2.name, E2.outputCompartmentName(), d1.name, d1.targetName())
+                self.W2.inputs << self.z1.s
+                self.z2.j << self.W2.outputs
+                self.e2.mu << self.z2.s
 
-        circuit.connect(z2.name, z2.outputCompartmentName(), e2.name, e2.meanName())#, bundle="additive")
-        #circuit.connect(zy.name, zy.outputCompartmentName(), e2.name, e2.targetName())
+                self.E2.inputs << self.e2.dmu
+                self.d1.target << self.E2.outputs
 
-        ## wire relevant compartment statistics to synaptic cable z0_z1
-        circuit.connect(z1.name, z1.surrogateCompartmentName(), d1.name, d1.modulatorName())
-        circuit.connect(z0.name, z0.outputCompartmentName(), W1.name, W1.presynapticCompartmentName())
-        circuit.connect(d1.name, d1.derivMeanName(), W1.name, W1.postsynapticCompartmentName())
-        circuit.connect(z1.name, z1.outputCompartmentName(), W2.name, W2.presynapticCompartmentName())
-        circuit.connect(e2.name, e2.derivMeanName(), W2.name, W2.postsynapticCompartmentName())
+                ## wire relevant compartment statistics to synaptic cable z0_z1
+                self.d1.modulator << self.z1.surrogate
+                self.W1.pre << self.z0.outputs
+                self.W1.post << self.d1.dmu
+                self.W2.pre << self.z1.s
+                self.W2.post << self.e2.dmu
 
-        ## make key commands known to model
-        circuit.add_command("reset", command_name="reset",
-                          component_names=[z0.name, W1.name, z1.name, W2.name, z2.name,
-                                           e2.name, E2.name, d1.name],
-                          reset_name="do_reset")
-        circuit.add_command("advance", command_name="advance",
-                          component_names=[z0.name, W1.name, z1.name, W2.name, z2.name,
-                                           e2.name, E2.name, d1.name])
-        circuit.add_command("evolve", command_name="evolve", component_names=[W1.name, W2.name])
-        circuit.add_command("clamp", command_name="clamp_input",
-                                 component_names=[z0.name], compartment=z0.inputCompartmentName(),
-                                 clamp_name="x")
-        circuit.add_command("clamp", command_name="clamp_target",
-                                 component_names=[e2.name], compartment=e2.targetName(),
-                                 clamp_name="y")
-        circuit.add_command("save", command_name="save", component_names=[W1.name, W2.name, E2.name,
-                                                                          z1.name, z2.name], directory_flag="dir")
+                reset_cmd, reset_args = self.circuit.compile_command_key(
+                                            self.z0, self.W1, self.z1,
+                                            self.W2, self.z2, self.e2,
+                                            self.E2, self.d1,
+                                            compile_key="reset")
+                advance_cmd, advance_args = self.circuit.compile_command_key(
+                                                self.z0, self.W1, self.z1,
+                                                self.W2, self.z2, self.e2,
+                                                self.E2, self.d1,
+                                                compile_key="advance_state")
+                evolve_cmd, evolve_args = self.circuit.compile_command_key(self.W1, self.W2, compile_key="evolve")
 
-        ## tell model the order in which to run automatic commands
-        # circuit.add_step("clamp_input")
-        circuit.add_step("advance")
-        #circuit.add_step("evolve")
-        ################################################################################
+                #self.circuit.add_command(wrap_command(jit(reset_cmd)), name="reset")
+                self.dynamic()
 
-        if save_init == True: ## save JSON structure to disk once
-            circuit.save_to_json(directory="exp", model_name=model_name)
-        self.model_dir = "{}/{}/custom".format(exp_dir, model_name)
-        if save_init == True:
-            circuit.save(dir=self.model_dir) ## save current parameter arrays
-        self.circuit = circuit # embed circuit to model construct
+            # reset, _ = reset.compile()
+            # self.reset = wrapper(jit(reset))
+            # advance, _ = advance.compile()
+            # self.advance = wrapper(jit(advance))
+            # evolve, _ = evolve.compile()
+            # self.evolve = wrapper(jit(evolve))
 
-    def save_to_disk(self):
+    def dynamic(self):## create dynamic commands for circuit
+        #from ngcsimlib.utils import get_current_context
+        #context = get_current_context()
+        z0, W1, z1, W2, z2, e2, E2, d1 = self.circuit.get_components("z0", "W1", "z1", "W2", "z2", "e2", "E2", "d1")
+        self.z0 = z0
+        self.W1 = W1
+        self.z1 = z1
+        self.W2 = W2
+        self.z2 = z2
+        self.e2 = e2
+        self.E2 = E2
+        self.d1 = d1
+        self.nodes = [z0, W1, z1, W2, z2, e2, E2, d1]
+
+        self.circuit.add_command(wrap_command(jit(self.circuit.reset)), name="reset")
+        self.circuit.add_command(wrap_command(jit(self.circuit.advance_state)), name="advance")
+        self.circuit.add_command(wrap_command(jit(self.circuit.evolve)), name="evolve")
+
+        # @Context.dynamicCommand
+        # def norm():
+        #     W1.weights.set(normalize_matrix(W1.weights.value, self.wNorm, order=1, axis=0))
+
+        @Context.dynamicCommand
+        def clamp(x, y):
+            z0.inputs.set(x)
+            e2.target.set(y)
+
+        @scanner
+        def process(compartment_values, args):
+            t = args[0]
+            dt = args[1]
+            compartment_values = self.circuit.advance_state(compartment_values, t, dt)
+            compartment_values = self.circuit.evolve(compartment_values, t, dt)
+            return compartment_values, (compartment_values[self.z1.s.path],
+                                        compartment_values[self.z2.s.path])
+
+
+    def save_to_disk(self, params_only=False):
         """
         Saves current model parameter values to disk
-        """
-        self.circuit.save(dir=self.model_dir) ## save current parameter arrays
 
-    def load_from_disk(self, model_directory="exp"):
+        Args:
+            params_only: if True, save only param arrays to disk (and not JSON sim/model structure)
+        """
+        if params_only == True:
+            model_dir = "{}/{}/custom".format(self.exp_dir, self.model_name)
+            self.W1.save(model_dir)
+            self.z1.save(model_dir)
+            self.W2.save(model_dir)
+            self.z2.save(model_dir)
+        else:
+            self.circuit.save_to_json(self.exp_dir, self.model_name) ## save current parameter arrays
+
+    def load_from_disk(self, model_directory):
         """
         Loads parameter/config values from disk to this model
 
         Args:
             model_directory: directory/path to saved model parameter/config values
         """
-        self.circuit.load_from_dir(self, model_directory)
+        with Context("Circuit") as circuit:
+            self.circuit = circuit
+            #self.circuit.load_from_dir(self.exp_dir + "/{}".format(self.model_name))
+            self.circuit.load_from_dir(model_directory)
+            ## note: redo scanner and anything using decorators
+            self.dynamic()
 
     def get_synapse_stats(self):
         """
@@ -219,8 +232,8 @@ class BFA_SNN():
         Returns:
             string containing min, max, mean, and L2 norm of W1
         """
-        _W1 = self.circuit.components.get("W1").weights
-        _W2 = self.circuit.components.get("W2").weights
+        _W1 = self.W1.weights.value
+        _W2 = self.W2.weights.value
         msg = "W1.n = {}  W2.n = {}".format(jnp.linalg.norm(_W1), jnp.linalg.norm(_W2))
         return msg
 
@@ -235,7 +248,7 @@ class BFA_SNN():
         Args:
             obs: observed pattern(s) to have spiking model process
 
-            lab: label pattern(s) to have spiking model use (used only if 
+            lab: label pattern(s) to have spiking model use (used only if
                 `adapt_synapses = True`)
 
             adapt_synapses: if True, synaptic efficacies will be adapted in
@@ -260,55 +273,95 @@ class BFA_SNN():
             an array containing spike vectors (will be empty; length = 0 if
                 collect_spike_train is False), estimated label distribution
         """
+        ## check and configure batch size
+        for node in self.nodes:
+            node.batch_size = obs.shape[0]
+
+        ## now run the model with configured batch size
         _obs = _scale(obs, input_gain)
         rGamma = 1.
         _S = []
         if get_latent_rates == True:
-            _S = jnp.zeros((obs.shape[0], self.circuit.components["z1"].n_units))
-        yMu = jnp.zeros((obs.shape[0], self.circuit.components["z2"].n_units))
+            # _S = jnp.zeros((obs.shape[0], self.circuit.components["z1"].n_units))
+            _S = jnp.zeros((obs.shape[0], self.z1.n_units))
+        # yMu = jnp.zeros((obs.shape[0], self.circuit.components["z2"].n_units))
+        yMu = jnp.zeros((obs.shape[0], self.z2.n_units))
         yCnt = yMu + 0
-        self.circuit.reset(do_reset=True)
+        #print(">> RESET START")
+        self.circuit.reset()
+        #print(">> RESET DONE")
         T_learn = 0.
         for ts in range(1, self.T):
-            self.circuit.clamp_input(_obs) #x=inp)
-            self.circuit.clamp_target(lab) #y=lab
-            self.circuit.runCycle(t=ts*self.dt, dt=self.dt)
+            #print(">> CLAMP START")
+            self.circuit.clamp(_obs, lab)
+            #print(">> CLAMP DONE")
+            #print(">> ADVANCE START")
+            self.circuit.advance(ts*self.dt, self.dt)
+            #print(">> ADVANCE DONE")
             curr_t = ts * self.dt ## track current time
+
             if adapt_synapses == True:
                 if curr_t > self.burnin_T:
-                    self.circuit.evolve(t=self.T, dt=self.dt)
-            yCnt = _add(self.circuit.components["z2"].spikes, yCnt)
+                    #print(">> ADVANCE DONE")
+                    self.circuit.evolve(ts*self.dt, self.dt)
+                    #print(">> EVOVLE DONE")
+            yCnt = _add(self.z2.s.value, yCnt)
+
             ## estimate output distribution
             if curr_t > self.burnin_T:
                 T_learn += 1.
                 if label_dist_estimator == "current":
-                    yMu = _add(self.circuit.components["z2"].current, yMu)
+                    yMu = _add(self.z2.j.value, yMu)
                 elif label_dist_estimator == "voltage":
-                    yMu = _add(self.circuit.components["z2"].voltage, yMu)
-                else: # label_dist_estimator == "spike":
-                    yMu = _add(self.circuit.components["z2"].spikes, yMu)
+                    yMu = _add(self.z2.v.value, yMu)
+                else:
+                    yMu = _add(self.z2.s.value, yMu)
             ## collect internal/hidden spikes at t
             if get_latent_rates == True:
-                _S = _add(_S, self.circuit.components["z1"].spikes)
+                _S = _add(_S, self.z1.s.value)
             else:
-                _S.append(self.circuit.components["z1"].spikes)
+                _S.append(self.z1.s.value)
 
-            ######### Logging/Model Matching #############
-            # if ts == 2:
-            #     print(f"{component_stats(self.circuit.components.get('z0'))}")
-            #     print(f"{component_stats(self.circuit.components.get('W1'))}")
-            #     print(f"{component_stats(self.circuit.components.get('z1'))}")
-            #     print(f"{component_stats(self.circuit.components.get('W2'))}")
-            #     print(f"{component_stats(self.circuit.components.get('z2'))}")
-            #     print(f"{component_stats(self.circuit.components.get('e2'))}")
-            #     print(f"{component_stats(self.circuit.components.get('E2'))}")
-            #     print(f"{component_stats(self.circuit.components.get('d1'))}")
-            #     sys.exit(0)
-            ##############################################
+        ## viet's old code
+        # _S = []
+        # if get_latent_rates == True:
+        #     # _S = jnp.zeros((obs.shape[0], self.circuit.components["z1"].n_units))
+        #     _S = jnp.zeros((obs.shape[0], self.z1.n_units))
+        # # yMu = jnp.zeros((obs.shape[0], self.circuit.components["z2"].n_units))
+        # yMu = jnp.zeros((obs.shape[0], self.z2.n_units))
+        # yCnt = yMu + 0
+        # self.reset()
+        # T_learn = 0.
+        # for ts in range(1, self.T):
+        #     # print(f"---- [TIME {ts}] ----")
+        #     self.z0.inputs.set(_obs)
+        #     self.e2.target.set(lab)
+        #     # print(f"[Step {ts}] z0.inputs: {self.z0.outputs.value.shape}, W1.outputs: {self.W1.outputs.value}, z1.s: {self.z1.s.value.shape}")
+        #     self.advance(ts*self.dt, self.dt)
+        #     curr_t = ts * self.dt ## track current time
+        #     if adapt_synapses == True:
+        #         if curr_t > self.burnin_T:
+        #             self.evolve(self.T, self.dt)
+        #     yCnt = _add(self.z2.s.value, yCnt)
+        #     ## estimate output distribution
+        #     if curr_t > self.burnin_T:
+        #         T_learn += 1.
+        #         if label_dist_estimator == "current":
+        #             yMu = _add(self.z2.j.value, yMu)
+        #         elif label_dist_estimator == "voltage":
+        #             yMu = _add(self.z2.v.value, yMu)
+        #         else:
+        #             yMu = _add(self.z2.s.value, yMu)
+        #     ## collect internal/hidden spikes at t
+        #     if get_latent_rates == True:
+        #         _S = _add(_S, self.z1.s.value)
+        #     else:
+        #         _S.append(self.z1.s.value)
 
         _yMu = softmax(yMu/T_learn) #self.T) ## estimate approximate label distribution
         if get_latent_rates == True:
             _S = (_S * rGamma)/self.T
+
         return _S, _yMu, yCnt
 
 @jit
