--- exhibits/bfa_snn/sim_new.sh
@@ -1,14 +1,0 @@
-#!/bin/sh
-
-################################################################################
-# Simulate the BFA-SNN on the MNIST database
-################################################################################
-
-DATA_DIR="../data/mnist"
-
-rm -r exp/* ## clear out experimental directory
-python train_bfasnn_new.py  --dataX="$DATA_DIR/trainX.npy" \
-                        --dataY="$DATA_DIR/trainY.npy" \
-                        --devX="$DATA_DIR/validX.npy" \
-                        --devY="$DATA_DIR/validY.npy" \
-                        --verbosity=1
